Move flyway version test into MigrateToModularStartersTest
